+++ include/HashSet.h
@@ -90,7 +90,7 @@ namespace Lucene
 		
 		virtual int32_t hashCode()
 		{
-			return (int64_t)setContainer.get();
+			return (int32_t)(int64_t)setContainer.get();
 		}
 		
 		operator bool() const
